적 hpBar 추가
1.Enemy.cs에 EnemyHp object 및 EnemyHp, EnemyMaxHp 필드 추가.
2.UI 캔버스 transform을 부모로 hpBar
UI 화면에 표시.
3. Enemy Spawner spawnTime 0.5 -> 1.0 변경
4.  EnemySpawenr.cs 에 적 hpBar List 추가 적 오브젝트 삭제시, 같이 삭제.

+++ Assets/Scripts/Projectile.cs
@@ -6,7 +6,7 @@ public class Projectile : MonoBehaviour
 {
     private Movement2D movement2d;
     private Transform target;
-    private int damage;
+    private float damage;
     private void Update()
     {
         if(target != null)
@@ -29,7 +29,7 @@ public class Projectile : MonoBehaviour
         collision.GetComponent<Enemy>().TakeDamage(damage);
     }
     
-    public void Setup(Transform target, int damage)
+    public void Setup(Transform target, float damage)
     {
         movement2d = GetComponent<Movement2D>();
         this.target = target;
